Use YAML config for submission; move submit scripts to production/; some reformatting
